WIP fixing complilation and syntax errors after updating the route line related classes to Carbon.

+++ librouteline/src/main/java/com/mapbox/navigation/ui/internal/route/RouteLayerProvider.kt
@@ -1,31 +1,32 @@
 package com.mapbox.navigation.ui.internal.route
 
 import android.graphics.drawable.Drawable
-import com.mapbox.mapboxsdk.maps.Style
-import com.mapbox.mapboxsdk.style.layers.LineLayer
-import com.mapbox.mapboxsdk.style.layers.SymbolLayer
+import com.mapbox.maps.Style
+import com.mapbox.maps.plugin.style.layers.LineLayer
+import com.mapbox.maps.plugin.style.layers.SymbolLayer
+
 
 interface RouteLayerProvider {
-    fun initializePrimaryRouteCasingLayer(style: Style, scale: Float, color: Int): LineLayer
-    fun initializeAlternativeRouteCasingLayer(style: Style, scale: Float, color: Int): LineLayer
+    fun initializePrimaryRouteCasingLayer(style: Style, scale: Double, color: Int): LineLayer
+    fun initializeAlternativeRouteCasingLayer(style: Style, scale: Double, color: Int): LineLayer
     fun initializePrimaryRouteLayer(
         style: Style,
         roundedLineCap: Boolean,
-        scale: Float,
+        scale: Double,
         color: Int
     ): LineLayer
 
     fun initializePrimaryRouteTrafficLayer(
         style: Style,
         roundedLineCap: Boolean,
-        scale: Float,
+        scale: Double,
         color: Int
     ): LineLayer
 
     fun initializeAlternativeRouteLayer(
         style: Style,
         roundedLineCap: Boolean,
-        scale: Float,
+        scale: Double,
         color: Int
     ): LineLayer
